+++ src/components/atoms/_well/index.js
@@ -1,19 +1,3 @@
-import React from 'react'
-import styled from 'styled-components'
-
-import { colors, spacing } from '@auth0/cosmos-tokens'
-
-const StyledWell = styled.div`
-  background: ${colors.form.background};
-  padding: ${spacing.large};
-`
-
-const Well = props => {
-  return <StyledWell {...props}>{props.children}</StyledWell>
-}
-
-Well.propTypes = {}
-
-Well.defaultProps = {}
+import Well from './well'
 
 export default Well
